requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
torch
matplotlib
opencv-python
